+++ packages/components/choropleth/src/projections.js
@@ -2,21 +2,12 @@ export {
 	// Azimuthal Projections
 	geoAzimuthalEqualArea,
 	geoAzimuthalEquidistant,
-	geoGnomonic,
-	geoOrthographic,
-	geoStereographic,
 
 	// Equal-Earth
 	geoEqualEarth,
 
-	// Conic Projections
-	geoConicConformal,
-	geoConicEqualArea,
-	geoConicEquidistant,
-
 	// Cylindrical Projections
 	geoEquirectangular,
 	geoMercator,
-	geoTransverseMercator,
 	geoNaturalEarth1
 } from 'd3-geo';