+++ packages/components/choropleth/CHANGELOG.md
@@ -1,3 +1,8 @@
+## `@svizzle/choropleth` v0.4.0 (next)
+
+- add `projectionId` so that we can pass either a `projection` prop (a projection function) or an id to select a projection from the available ones
+- remove more exotic projections that would better be used with spherical coordinates [lambda, phi]
+
 ## `@svizzle/choropleth` v0.3.0
 
 - `ChoroplethG`